+++ packages/uba/CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - characters `,.[]{}` added to password policy complexity check.
+ Checking is carried out in case ubs_settings `UBA.passwordPolicy.checkCmplexity` value is set to `true`.  
 
 ## [5.4.7] - 2020-04-24
 ## [5.4.6] - 2020-04-10