+++ packages/uba/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - **SECURITY** access to `uba_user.uPasswordHashHexa` is disallowed from client (restricted for all)
  - **SECURITY** access to `uba_user` `disabled`, `isPending` and `trustedIP` attributes are allowed only for members of 
  `Supervisor` role; For other roles actual values is replaced by `*****` for trustedIP and `null` for `disabled`, `isPending`
- - SMS registration handled is added to `uba_user.publicRegistration` rest method. Registration kind can be defined
+ - SMS registration handler added to `uba_user.publicRegistration` rest method. Registration kind can be defined
   in `serverConfig.application.customSettings.publicRegistration.kind` application config key.
   See `uba_user.publicRegistration` documentation for details.